Conflicts:
	src/modules/rrhh/RegistroPersonal/components/Dialogos/RegistroPersonalDialog.vue

+++ src/modules/rrhh/RegistroPersonal/components/RegistroPersonal.vue
@@ -1,5 +1,5 @@
 <template>
-  <h3 class="text-center mb-4">Mantenimientos Empleados</h3>
+  <h3 class="text-center mb-4">Registro de Empleados</h3>
 
   <CCard class="mb-4">
     <CCardBody class="table-headers justify-content-between">
@@ -59,16 +59,23 @@
     @closeModal="closeTipoNovedad"
   />
   
-  <ContenedorArchivosRRHH
+  <ContenedorArchivosModel
     :showModal="showModalDoc"
-    :empleado="selectedEmployee"
-    @custom-event="closeContenedorModal"
+    :tagKeyName="'empleadoId'"
+    :tagValueName="selectedEmployee?.id"
+    @closeModal="closeContenedorModal"
   />
 
   <EducacionDialog
     :showModal="showEducacion"
     :employeeInfo="selectedEmployee"
     @closeModal="() => showEducacion = false"
+  />
+
+  <UtilesLaboralesDialog
+    :showModal="showUtilesLaboralesDialog"
+    :employeeInfo="selectedEmployee"
+    @closeModal="() => showUtilesLaboralesDialog = false"
   />
 
   <!-- Reportes -->
@@ -102,10 +109,13 @@ import { useToastStore } from '@/store/toast'
 import RegistroPersonalDialog from '../components/Dialogos/RegistroPersonalDialog.vue'
 import RegistroPersonalTable from '../components/tables/RegistroPersonalTable.vue'
 import EducacionDialog from './Dialogos/EducacionDialog.vue'
-import ContenedorArchivosRRHH from './ContenedorArchivosRRHH.vue'
+import ContenedorArchivosModel from '@/components/ContenedorArchivosModel.vue'
 import AccionPersonalDialog from './Dialogos/AccionPersonalDialog.vue'
+import UtilesLaboralesDialog from './Dialogos/UtilesLaboralesDialog.vue'
 import TipoNovedadDialog from './TipoNovedades.vue'
 import TarjetaEmpleadoDialogs from '../components/Dialogos/TarjetaEmpleado.vue'
+import Swal from 'sweetalert2/dist/sweetalert2.js'
+import { showReport } from '@/utils/util'
 
 export default {
   components: {
@@ -114,12 +124,13 @@ export default {
     RegistroPersonalDialog,
     AccionPersonalDialog,
     AccionPersonalDialog,
-    ContenedorArchivosRRHH,
+    ContenedorArchivosModel,
     TipoNovedadDialog,
     RegistroPersonalDialog,
     TarjetaEmpleadoDialogs,
     EmpleadoReports,
     EducacionDialog,
+    UtilesLaboralesDialog,
   },
 
   data: function () {
@@ -136,13 +147,14 @@ export default {
       showEducacion: false,
       employeeInfo: null,
       showRegistroPersonalModal: false,
+      showUtilesLaboralesDialog: false,
       registroPersonal: [],
       reporteDepto: '1',
       reportes: false,
       columns: [
         { key: 'codigo', label: 'Código', _style: { width: '5%' } },
-        { key: 'apellidos', label: 'Apellido', _style: { width: '15%' } },
-        { key: 'nombres', label: 'Nombre', _style: { width: '15%' } },
+        { key: 'apellido', label: 'Apellido', _style: { width: '15%' } },
+        { key: 'nombre', label: 'Nombre', _style: { width: '15%' } },
         { key: 'codigoIdentidad', label: 'Cédula/Pasaporte', _style: { width: '10%' } },
         {
           key: 'departamenteName',
@@ -196,14 +208,21 @@ export default {
         {
           label: 'Acción personal',
           clickHandler: (value) => {
+            this.selectedEmployee = value
             this.showAccionPersonal = true
+          }
+        },
+        {
+          label: 'Tabla de acciones',
+          clickHandler: (value) => {
+            this.showTipoNovedad = true
             this.selectedEmployee = value
           }
         },
         {
-          label: 'Eventualidad',
+          label: 'Útiles laborales',
           clickHandler: (value) => {
-            this.showTipoNovedad = true
+            this.showUtilesLaboralesDialog = true
             this.selectedEmployee = value
           }
         },
@@ -255,42 +274,42 @@ export default {
       this.showRegistroPersonalModal = true
     },
 
-    imprimirReporte() {
-      if (this.reporteDepto === '1') {
-        window
-          .open(
-            `http://lmd-server-01/ReportServer/Pages/ReportViewer.aspx?%2fRRHH%2fRep_Empleados_por_Nombre&rs:Command=Render&ID_AYUNTAMIENTO=${this.$ayuntamientoId}`,
-            '_blank',
-          )
-          .focus()
-      }
-      
-      if (this.reporteDepto === '2') {
-        window
-          .open(
-            `http://lmd-server-01/ReportServer/Pages/ReportViewer.aspx?%2fRRHH%2fRep_Empleados_por_Apellidos&rs:Command=Render&ID_AYUNTAMIENTO=${this.$ayuntamientoId}`,
-            '_blank',
-          )
-          .focus()
-      }
-      if (this.reporteDepto === '3') {
-        window
-          .open(
-            `http://lmd-server-01/ReportServer/Pages/ReportViewer.aspx?%2fRRHH%2fRep_Empleados_por_Cargo&rs:Command=Render&ID_AYUNTAMIENTO=${this.$ayuntamientoId}`,
-            '_blank',
-          )
-          .focus()
-      }
-      if (this.reporteDepto === '4') {
-        window
-          .open(
-            `http://lmd-server-01/ReportServer/Pages/ReportViewer.aspx?%2fRRHH%2fRep_Departamentos&rs:Command=Render&ID_AYUNTAMIENTO=${this.$ayuntamientoId}`,
-            '_blank',
-          )
-          .focus()
-      }
+    async imprimirReporte() {
+      const reportParam = {
+        folderName: 'rrhh',
+        params: [{
+          name: "ID_AYUNTAMIENTO",
+          value: "majorityId"
+        }]
+      };
 
-      this.reportes = false;
+      try {
+        if (this.reporteDepto === '1') {
+          reportParam.reportName = 'Rep_Empleados_por_Nombre';
+          await showReport(reportParam);
+        }
+        if (this.reporteDepto === '2') {
+          reportParam.reportName = 'Rep_Empleados_por_Apellidos';
+          await showReport(reportParam);
+        }
+        if (this.reporteDepto === '3') {
+          reportParam.reportName = 'Rep_Empleados_por_Cargo';
+          await showReport(reportParam);
+        }
+        if (this.reporteDepto === '4') {
+          reportParam.reportName = 'Rep_Empleados_por_Departamento';
+          await showReport(reportParam);
+        }
+
+        this.reportes = false;
+      } catch (error) {
+        this.show({
+          content: error,
+          closable: true,
+          color: 'danger',
+          class: 'text-white',
+        })
+      }
     },
 
     closeContenedorModal(payload) {
@@ -305,8 +324,8 @@ export default {
     },
 
     toggleDetails(item) {
-      this.showModal();
       this.selectedEmployee = item;
+      this.showModal();
     },
 
     submitForm(payload) {
@@ -347,23 +366,38 @@ export default {
     },
 
     deleteEmp(item) {
-      Api.deleteEmpleado(item.id)
-        .then((response) => {
-          this.show({
-            content: response.data.message,
-            closable: true,
-            color: 'inherit',
+      Swal.fire({
+        position: 'center',
+        icon: 'warning',
+        title: `Estás usted seguro que quieres eliminar este empleado?`,
+        showConfirmButton: true,
+        confirmButtonText: 'Si',
+        cancelButtonText: 'No',
+        showCancelButton: true,
+        allowEscapeKey: false,
+        allowOutsideClick: false,
+        customClass: 'btns',
+      }).then((answer) => {
+        if (answer.isConfirmed) {
+          Api.deleteEmpleado(item.id)
+          .then((response) => {
+            this.show({
+              content: response.data.message,
+              closable: true,
+              color: 'inherit',
+            })
+            setTimeout(this.getRegistroPersonal, 500)
           })
-          setTimeout(this.getRegistroPersonal, 500)
-        })
-        .catch((error) => {
-          this.show({
-            content: error.response.data,
-            closable: true,
-            color: 'danger',
-            class: 'text-white',
+          .catch((error) => {
+            this.show({
+              content: error.response.data,
+              closable: true,
+              color: 'danger',
+              class: 'text-white',
+            })
           })
-        })
+        }
+      })
     },
 
     handleFilterEmployeeByStatus({ target }) {
